+++ gpflow/covariances/__init__.py
@@ -1,4 +1,5 @@
 from .dispatch import Kuf
 from .dispatch import Kuu
 
-from . import kufs, kuus, mo_kufs, mo_kuus
+from . import kufs, kuus
+from .multioutput import kufs as mo_kufs, kuus as mo_kuus
