+++ gpflow/inducing_variables/__init__.py
@@ -1,6 +1,7 @@
 from .inducing_variables import InducingVariables, InducingPoints, Multiscale
 from .inducing_patch import InducingPatches
-from .mo_inducing_variables import (
+from .multioutput import inducing_variables as mo_inducing_variables
+from .multioutput.inducing_variables import (
     MultioutputInducingVariables,
     FallbackSharedIndependentInducingVariables,
     FallbackSeparateIndependentInducingVariables,
