+++ gpflow/kernels/__init__.py
@@ -3,7 +3,7 @@ from .convolutional import Convolutional
 from .changepoints import ChangePoints
 from .linears import Linear, Polynomial
 from .misc import ArcCosine, Coregion
-from .mo_kernels import (
+from .multioutput.kernels import (
     MultioutputKernel,
     SeparateIndependent,
     SharedIndependent,
